+++ plugin-dev/Source/Sentry/Private/Tests/SentryEvent.spec.cpp
@@ -34,6 +34,33 @@ void SentryEventSpec::Define()
 			TestFalse("Event ID is non-empty", SentryEvent->GetId().IsEmpty());
 		});
 	});
+
+	Describe("Event fingerprint", [this]()
+	{
+		It("should persist its value", [this]()
+		{
+			TArray<FString> InFingerprint = { TEXT("F1"), TEXT("F2"), TEXT("F3") };
+
+			SentryEvent->SetFingerprint(InFingerprint);
+			TArray<FString> OutFingerprint = SentryEvent->GetFingerprint();
+
+			TestEqual("Fingerprint elements count", OutFingerprint.Num(), InFingerprint.Num());
+			TestEqual("Fingerprint first element", OutFingerprint[0], InFingerprint[0]);
+			TestEqual("Fingerprint second element", OutFingerprint[1], InFingerprint[1]);
+			TestEqual("Fingerprint third element", OutFingerprint[2], InFingerprint[2]);
+		});
+
+		It("can be emptry", [this]()
+		{
+			TArray<FString> InFingerprint = {};
+
+			SentryEvent->SetFingerprint(InFingerprint);
+			TArray<FString> OutFingerprint = SentryEvent->GetFingerprint();
+
+			TestEqual("Fingerprint elements count", OutFingerprint.Num(), InFingerprint.Num());
+			TestEqual("Fingerprint is empty", OutFingerprint.Num(), 0);
+		});
+	});
 }
 
 #endif